requirements.txt
requests
redis
python-dotenv
pytz
asyncio
cogwatch
sqlalchemy
mariadb
py-cord
Flask>=2.0
Flask-Session
gunicorn>=20.0
gevent